### 2020-03-10 #2156 @danielskatz
👋 @JonasRieger - thanks for your submission.  I will now find an editor for it.  While I do so, can you:

1. fix the cases in the references, using {}s to protect cases, along with any other proofreading of the paper. When you have made changes, you can add a new comment here with ```@whedon generate pdf``` to see the built pdf.  One example is "tools for r" which should be "tools for R"

1. Suggest potential reviewers - please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
