+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @Manual{R,
-    title = {R: A Language and Environment for Statistical Computing},
+    title = {R: {A} {Language} and {Environment} for {Statistical} {Computing}},
     author = {{R Core Team}},
     organization = {R Foundation for Statistical Computing},
     address = {Vienna, Austria},
@@ -8,7 +8,7 @@
   }
 
 @Manual{tosca,
-    title = {tosca: Tools for Statistical Content Analysis},
+    title = {tosca: {Tools} for {Statistical} {Content} {Analysis}},
     author = {Lars Koppers and Jonas Rieger and Karin Boczek and Gerret {von Nordheim}},
     year = {2019},
     url = {https://github.com/Docma-TU/tosca},
@@ -17,7 +17,7 @@
   }
 
 @Article{batchtools,
-    title = {batchtools: Tools for R to work on batch systems},
+    title = {batchtools: Tools for {R} to work on batch systems},
     author = {Michel Lang and Bernd Bischl and Dirk Surmann},
     journal = {The Journal of Open Source Software},
     year = {2017},
@@ -28,7 +28,7 @@
   }
   
   @Manual{parallelMap,
-    title = {parallelMap: Unified Interface to Parallelization Back-Ends},
+    title = {parallelMap: {Unified} {Interface} to {Parallelization} {Back-Ends}},
     author = {Bernd Bischl and Michel Lang},
     year = {2019},
     note = {R package version 1.4},
@@ -36,7 +36,7 @@
   }
   
   @Manual{lda,
-    title = {lda: Collapsed Gibbs Sampling Methods for Topic Models},
+    title = {lda: Collapsed {Gibbs} {Sampling} {Methods} for {Topic} {Models}},
     author = {Jonathan Chang},
     year = {2015},
     note = {R package version 1.4.2},
@@ -44,7 +44,7 @@
   }
   
   @Article{topicmodels,
-    title = {{topicmodels}: An {R} Package for Fitting Topic Models},
+    title = {{topicmodels}: An {R} {Package} for {Fitting} {Topic} {Models}},
     author = {Bettina Gr\"un and Kurt Hornik},
     journal = {Journal of Statistical Software},
     year = {2011},
@@ -69,7 +69,7 @@
 
 @article{blei2003,
  author = {Blei, David M. and Ng, Andrew Y. and Jordan, Michael I.},
- title = {Latent Dirichlet Allocation},
+ title = {Latent {Dirichlet} {Allocation}},
  journal = {Journal of Machine Learning Research},
  issue_date = {3/1/2003},
  volume = {3},
@@ -84,7 +84,7 @@
 
 @article{blei2012,
  author = {Blei, David M.},
- title = {Probabilistic Topic Models},
+ title = {Probabilistic {Topic} {Models}},
  journal = {Communications of the ACM},
  issue_date = {April 2012},
  volume = {55},
@@ -101,7 +101,7 @@
 
 @inproceedings{chang2009,
  author = {Chang, Jonathan and Boyd-Graber, Jordan and Gerrish, Sean and Wang, Chong and Blei, David M.},
- title = {Reading Tea Leaves: How Humans Interpret Topic Models},
+ title = {Reading {Tea} {Leaves}: How {Humans} {Interpret} {Topic} {Models}},
  year = {2009},
  isbn = {9781615679119},
  publisher = {Curran Associates Inc.},
