+++ paper/paper.bib
@@ -8,7 +8,7 @@
   }
 
 @Manual{tosca,
-    title = {tosca: {Tools} for {Statistical} {Content} {Analysis}},
+    title = {{tosca}: {Tools} for {Statistical} {Content} {Analysis}},
     author = {Lars Koppers and Jonas Rieger and Karin Boczek and Gerret {von Nordheim}},
     year = {2019},
     url = {https://github.com/Docma-TU/tosca},
@@ -17,7 +17,7 @@
   }
 
 @Article{batchtools,
-    title = {batchtools: Tools for {R} to work on batch systems},
+    title = {{batchtools}: Tools for {R} to work on batch systems},
     author = {Michel Lang and Bernd Bischl and Dirk Surmann},
     journal = {The Journal of Open Source Software},
     year = {2017},
@@ -28,7 +28,7 @@
   }
   
   @Manual{parallelMap,
-    title = {parallelMap: {Unified} {Interface} to {Parallelization} {Back-Ends}},
+    title = {{parallelMap}: {Unified} {Interface} to {Parallelization} {Back-Ends}},
     author = {Bernd Bischl and Michel Lang},
     year = {2019},
     note = {R package version 1.4},
@@ -36,7 +36,7 @@
   }
   
   @Manual{lda,
-    title = {lda: Collapsed {Gibbs} {Sampling} {Methods} for {Topic} {Models}},
+    title = {{lda}: Collapsed {Gibbs} {Sampling} {Methods} for {Topic} {Models}},
     author = {Jonathan Chang},
     year = {2015},
     note = {R package version 1.4.2},
