### 2020-05-03 #2181 @whedon
[REVIEW]: ldaPrototype: A method in R to get a Prototype of multiple Latent Dirichlet Allocations
**Submitting author:** @JonasRieger (<a href="http://orcid.org/0000-0002-0007-4478">Jonas Rieger</a>)
**Repository:** <a href="https://github.com/JonasRieger/ldaPrototype" target ="_blank">https://github.com/JonasRieger/ldaPrototype</a>
**Version:** v0.2.0
