### 2020-05-03 #2181 @TommyJones
Thanks @arfon!

The paper and software look great, but there are a couple issues, one is a bit of a blocker to me finishing the review.

The big one:
* There is no example usage of how to use the software for an analysis problem. A great place to put this would be in the README, showing basic usage. If you want to cover more ground than one would typically put in a README, a vignette is a good place. But without this, I'm not sure where to start and thus can't check functionality.

Minor issues:
* The community guidelines do not have a clear statement or instructions for contributing. Adding a sentence to the bottom of the README would fix that right up.
* Functionality documentation is "OK" according to the review criteria. Not all functions have examples, though many do. (Note: as I see it, it's still a "pass" but want to flag that it could be improved.)

One thing that is out of scope for JOSS but I think is worth mentioning: the documentation for parameter inputs to functions is clear but non-standard for R packages. Comes off as a little odd.
