### 2020-05-03 JonasRieger/ldaPrototype#1 @TommyJones
Add example analysis to README or Vignette
Relating to JOSS review [here](https://github.com/openjournals/joss-reviews/issues/2181#issuecomment-623142268)

There is no example usage of how to use the software for an analysis problem. A great place to put this would be in the README, showing basic usage. If you want to cover more ground than one would typically put in a README, a vignette is a good place. But without this, I'm not sure where to start and thus can't check functionality.
