### 2020-05-03 JonasRieger/ldaPrototype#2 @TommyJones
Contribution guidelines (as opposed to code of conduct) are unclear
Related to JOSS review [here](https://github.com/openjournals/joss-reviews/issues/2181#issuecomment-623142268)

The community guidelines do not have a clear statement or instructions for contributing. Adding a sentence to the bottom of the README would fix that right up.
