### 2020-05-03 #2181 @TommyJones
I've opened two issues in the target repo:

[This](https://github.com/JonasRieger/ldaPrototype/issues/1#issue-611517976) is for an example analysis to README or vignette.

[This](https://github.com/JonasRieger/ldaPrototype/issues/2#issue-611518164) is for contribution guidelines.
