### 2020-05-04 JonasRieger/ldaPrototype#1 @JonasRieger
Thanks for the comment. I fixed it. https://github.com/JonasRieger/ldaPrototype/commit/7b17307cb5b760650d23e03aa8cba6590c0e996d
