### 2020-05-04 #2181 @JonasRieger
Thanks for the comments. I have improved the points. I will also look again where I can add examples in the documentation of single functions. I deliberately chose this style for the parameter documentation because it is clearer than the standard in R (which is mostly very rudimentary and merely prosaic). It is inspired by the packages [mlr](https://github.com/mlr-org/mlr) and [batchtools](https://github.com/mllg/batchtools).
