+++ R/LDABatch.R
@@ -47,7 +47,7 @@
 #' getJob(batch)
 #' getLDA(batch, 2)
 #'
-#' batch2 = LDABatch(docs = reuters_docs, vocab = reuters_vocab, K = 15, chunk.size = 20)
+#' batch2 = LDABatch(id = "Ex2", docs = reuters_docs, vocab = reuters_vocab, K = 15, chunk.size = 20)
 #' batch2
 #' head(getJob(batch2))
 #' }
