### 2020-05-04 cf3de36
donttest -> dontrun for LDABatch calls
