+++ man/as.LDABatch.Rd
@@ -41,7 +41,7 @@ a \code{\link{LDABatch}} object, which can be handled using the getter functions
 at \code{\link{getJob}}.
 }
 \examples{
-\donttest{
+\dontrun{
 batch = LDABatch(docs = reuters_docs, vocab = reuters_vocab, K = 15, chunk.size = 20)
 batch
 
