+++ man/as.LDARep.Rd
@@ -62,7 +62,7 @@ res2
 getJob(res2)
 getJob(res)
 
-\donttest{
+\dontrun{
 batch = LDABatch(docs = reuters_docs, vocab = reuters_vocab, n = 4, id = "TEMP", K = 30)
 res3 = as.LDARep(batch)
 res3
