### 2020-05-05 #2181 @TommyJones
@arfon, @karthik, @JonasRieger I've completed my review. Should this close out the issue?

Looks like a useful package! Opened up another issue in the target repo that I think will make for a more pleasant user experience, but think it falls outside of scope for JOSS as I understand it.
