### 2020-05-05 7838aeb
parallelMap, future, batchtools to imports
