### 2020-05-12 #2156 @bstewart
Sure- I will likely be a bit slower and if someone could direct me to the guidelines/expectations I'd be grateful.  But happy to help out.
