### 2020-05-15 #2181 @JonasRieger
@arfon @karthik In case you missed it, I would like to mention that a second reviewer was found in https://github.com/openjournals/joss-reviews/issues/2156.
