### 2020-05-15 #2181 @arfon
:wave:  @bstewart - many thanks for agreeing to review this submission - I've just helped out @karthik by adding your checklist to the top of this issue thread.

Please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Any questions/concerns please let @karthik or myself know.
