### 2020-05-15 68fef10
new similarity measures
