+++ R/jaccardTopics.R
@@ -58,8 +58,9 @@
 #'   words for similarity calculation. Could differ from \code{wordslimit}, if
 #'   \code{atLeast} is greater than zero.}
 #'   \item{\code{param}}{[\code{named list}] with parameter specifications for
+#'   \code{type} [\code{character(1)}] = "Jaccard Coefficient",
 #'   \code{limit.rel} [0,1], \code{limit.abs} [\code{integer(1)}] and
-#'   \code{atLeast} [\code{integer(1)}] See above for explanation.}
+#'   \code{atLeast} [\code{integer(1)}]. See above for explanation.}
 #' }
 #'
 #' @examples
@@ -74,6 +75,13 @@
 #'
 #' sim = getSimilarity(jacc)
 #' dim(sim)
+#'
+#' # Comparison to Cosine and Jensen-Shannon (more interesting on large datasets)
+#' cosine = cosineTopics(topics)
+#' js = jsTopics(topics)
+#'
+#' sims = list(jaccard = sim, cosine = getSimilarity(cosine), js = getSimilarity(js))
+#' pairs(do.call(cbind, lapply(sims, as.vector)))
 #'
 #' @export jaccardTopics
 
@@ -184,7 +192,8 @@ jaccardTopics.serial = function(topics, limit.rel, limit.abs, atLeast, progress 
   sims[is.nan(sims)] = 0
 
   res = list(sims = sims, wordslimit = wordsconsidered, wordsconsidered = colSums(index),
-    param = list(limit.rel = limit.rel, limit.abs = limit.abs, atLeast = atLeast))
+    param = list(type = "Jaccard Coefficient",
+      limit.rel = limit.rel, limit.abs = limit.abs, atLeast = atLeast))
   class(res) = "TopicSimilarity"
   res
 }
