+++ man/jsTopics.Rd
@@ -1,0 +1,86 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/jsTopics.R
+\name{jsTopics}
+\alias{jsTopics}
+\title{Pairwise Jensen-Shannon Similarities (Divergences)}
+\usage{
+jsTopics(topics, epsilon = 1e-06, progress = TRUE)
+}
+\arguments{
+\item{topics}{[\code{named matrix}]\cr
+The counts of vocabularies/words (row wise) in topics (column wise).}
+
+\item{epsilon}{[\code{numeric(1)}]\cr
+TODO!
+Default is \code{1e-6}.}
+
+\item{progress}{[\code{logical(1)}]\cr
+Should a nice progress bar be shown? Turning it off, could lead to significantly
+faster calculation. Default is \code{TRUE}.
+If \code{pm.backend} is set, parallelization is done and no progress bar will be shown.}
+
+\item{pm.backend}{[\code{character(1)}]\cr
+One of "multicore", "socket" or "mpi".
+If \code{pm.backend} is set, \code{\link[parallelMap]{parallelStart}} is
+called before computation is started and \code{\link[parallelMap]{parallelStop}}
+is called after.}
+
+\item{ncpus}{[\code{integer(1)}]\cr
+Number of (physical) CPUs to use. If \code{pm.backend} is passed,
+default is determined by \code{\link[future]{availableCores}}.}
+}
+\value{
+[\code{named list}] with entries
+\describe{
+  \item{\code{sims}}{[\code{lower triangular named matrix}] with all pairwise
+  similarities of the given topics.}
+  \item{\code{wordslimit}}{[\code{integer}] = vocabulary size. See
+  \code{\link{jaccardTopics}} for original purpose.}
+  \item{\code{wordsconsidered}}{[\code{integer}] = vocabulary size. See
+  \code{\link{jaccardTopics}} for original purpose.}
+  \item{\code{param}}{[\code{named list}] with parameter specifications for
+  \code{type} [\code{character(1)}] = "Cosine Similarity" and
+  \code{epsilon} [\code{numeric(1)}]. See above for explanation.}
+}
+}
+\description{
+Calculates the similarity of all pairwise topic combinations using the
+Jensen-Shannon Divergence.
+}
+\details{
+The Jensen-Shannon Similarity ...
+
+\eqn{J_m} is calculated by
+\deqn{J_m(\bm z_{i}, \bm z_{j} \mid \bm c) = \frac{\sum_{v = 1}^{V} 1_{\left\{n_{i}^{(v)} > c_i ~\wedge~ n_{j}^{(v)} > c_j\right\}}\left(n_{i}^{(v)}, n_{j}^{(v)}\right)}{\sum_{v = 1}^{V} 1_{\left\{n_{i}^{(v)} > c_i ~\vee~ n_{j}^{(v)} > c_j\right\}}\left(n_{i}^{(v)}, n_{j}^{(v)}\right)}}
+with \eqn{V} is the vocabulary size, and \eqn{n_k^{(v)}} is the count of
+assignments of the \eqn{v}-th word to the \eqn{k}-th topic. The threshold vector \eqn{\bm c}
+is determined by the maximum threshold of the user given lower bounds \code{limit.rel}
+and \code{limit.abs}. In addition, at least \code{atLeast} words per topic are
+considered for calculation. According to this, if there are less than
+\code{atLeast} words considered as relevant after applying \code{limit.rel}
+and \code{limit.abs} the \code{atLeast} most common words per topic are taken
+to determine topic similarities.
+}
[…]

[…]
