### 2020-05-15 b1bd2ab
fix test jaccard
