+++ R/jaccardTopics.R
@@ -58,7 +58,7 @@
 #'   words for similarity calculation. Could differ from \code{wordslimit}, if
 #'   \code{atLeast} is greater than zero.}
 #'   \item{\code{param}}{[\code{named list}] with parameter specifications for
-#'   \code{type} [\code{character(1)}] = "Jaccard Coefficient",
+#'   \code{type} [\code{character(1)}] \code{= "Jaccard Coefficient"},
 #'   \code{limit.rel} [0,1], \code{limit.abs} [\code{integer(1)}] and
 #'   \code{atLeast} [\code{integer(1)}]. See above for explanation.}
 #' }
@@ -159,7 +159,8 @@ jaccardTopics.parallel = function(topics, limit.rel, limit.abs, atLeast, pm.back
   sims[is.nan(sims)] = 0
 
   res = list(sims = sims, wordslimit = wordsconsidered, wordsconsidered = colSums(index),
-    param = list(limit.rel = limit.rel, limit.abs = limit.abs, atLeast = atLeast))
+    param = list(type = "Jaccard Coefficient",
+      limit.rel = limit.rel, limit.abs = limit.abs, atLeast = atLeast))
   class(res) = "TopicSimilarity"
   res
 }
