+++ R/jsTopics.R
@@ -46,7 +46,7 @@
 #'   \item{\code{wordsconsidered}}{[\code{integer}] = vocabulary size. See
 #'   \code{\link{jaccardTopics}} for original purpose.}
 #'   \item{\code{param}}{[\code{named list}] with parameter specifications for
-#'   \code{type} [\code{character(1)}] = "Cosine Similarity" and
+#'   \code{type} [\code{character(1)}] \code{= "Cosine Similarity"} and
 #'   \code{epsilon} [\code{numeric(1)}]. See above for explanation.}
 #' }
 #'
