+++ man/cosineTopics.Rd
@@ -35,7 +35,7 @@ default is determined by \code{\link[future]{availableCores}}.}
   \item{\code{wordsconsidered}}{[\code{integer}] = vocabulary size. See
   \code{\link{jaccardTopics}} for original purpose.}
   \item{\code{param}}{[\code{named list}] with parameter
-  \code{type} [\code{character(1)}] = "Cosine Similarity".}
+  \code{type} [\code{character(1)}] \code{= "Cosine Similarity"}.}
 }
 }
 \description{
