+++ man/getPrototype.Rd
@@ -131,8 +131,9 @@ Not considered for \code{\link{LDABatch}} or \code{\link{LDARep}} objects.}
   \item{\code{protoid}}{[\code{character(1)}] Name (ID) of the determined Prototype LDA.}
   \item{\code{id}}{[\code{character(1)}] See above.}
   \item{\code{param}}{[\code{named list}] with parameter specifications for
+  \code{type} [\code{character(1)}] \code{= "Jaccard Coefficient"},
   \code{limit.rel} [0,1], \code{limit.abs} [\code{integer(1)}] and
-  \code{atLeast} [\code{integer(1)}] See above for explanation.}
+  \code{atLeast} [\code{integer(1)}]. See above for explanation.}
   \item{\code{topics}}{[\code{named matrix}] with the count of vocabularies
   (row wise) in topics (column wise).}
   \item{\code{sims}}{[\code{lower triangular named matrix}] with all pairwise
