### 2020-05-15 2f26b9c
parallel cosine jensen shannon
