### 2020-05-15 3f7eab4
equation for bm
