+++ man/jsTopics.Rd
@@ -48,18 +48,12 @@ Calculates the similarity of all pairwise topic combinations using the
 Jensen-Shannon Divergence.
 }
 \details{
-The Jensen-Shannon Similarity for two topics \eqn{\bm z_{i}} and \eqn{\bm z_{j}} ...
+The Jensen-Shannon Similarity for two topics \eqn{\bm z_{i}} and
+\eqn{\bm z_{j}} is calculated by
+\deqn{\text{JS}(\bm z_{i}, \bm z_{j}) = 1 - \left( \text{KLD}(p_i, (p_i+p_j)/2) + \text{KLD}((p_j, (p_i+p_j)/2)) \right)/2 }
 
-\eqn{J_m} is calculated by
-\deqn{J_m(\bm z_{i}, \bm z_{j} \mid \bm c) = \frac{\sum_{v = 1}^{V} 1_{\left\{n_{i}^{(v)} > c_i ~\wedge~ n_{j}^{(v)} > c_j\right\}}\left(n_{i}^{(v)}, n_{j}^{(v)}\right)}{\sum_{v = 1}^{V} 1_{\left\{n_{i}^{(v)} > c_i ~\vee~ n_{j}^{(v)} > c_j\right\}}\left(n_{i}^{(v)}, n_{j}^{(v)}\right)}}
-with \eqn{V} is the vocabulary size, and \eqn{n_k^{(v)}} is the count of
-assignments of the \eqn{v}-th word to the \eqn{k}-th topic. The threshold vector \eqn{\bm c}
-is determined by the maximum threshold of the user given lower bounds \code{limit.rel}
-and \code{limit.abs}. In addition, at least \code{atLeast} words per topic are
-considered for calculation. According to this, if there are less than
-\code{atLeast} words considered as relevant after applying \code{limit.rel}
-and \code{limit.abs} the \code{atLeast} most common words per topic are taken
-to determine topic similarities.
+with \eqn{V} is the vocabulary size and \eqn{n_k^{(v)}} is the count of
+assignments of the \eqn{v}-th word to the \eqn{k}-th topic.
 }
 \examples{
 res = LDARep(docs = reuters_docs, vocab = reuters_vocab, n = 4, K = 10, num.iterations = 30)

[…]
