### 2020-06-17 JonasRieger/ldaPrototype#4 @bstewart
Readme Update
Following the code in the ReadMe breaks down at Step 3.1 because `sims` isn't an object that has been completed.  (by the way, I enjoyed the ReadMe it was nice that you highlighted the aggregate function and then broke down the components)
