### 2020-06-17 JonasRieger/ldaPrototype#5 @bstewart
Paper Suggestions
In reference to the JOSS Review, a few paper suggestions
- `mallet` is another package for estimating lda that might be mentioned along with `lda` and `topicmodels`.
- Regarding the line: "A large part of the analysis is based on this model (LDA).” I don’t know what ’the analysis’ is referencing.
- Regarding the line "Up to now, the so-called eye-balling method has been used in practice to select suitable results. From a set of models, subjective decisions are made to select the model that seems to fit the data best. This contradicts basically good scientific practice.” I think the more prominent technique has been selection by log-likelihood.  I also don’t think your characterization of good scientific practice is uncontroversial here—there are tons of decisions which are essentially subjective decisions (includes of decisions of how to do measurement, what to investigate, or even whether to use LDA at all).  It seems that someone who selects a model based on careful reading or fit for a particular/question of analysis would fall into bad scientific practice for your definition which feels unfair.  I’ll note that many of the choices in your design are equally arbitrary (e.g. the similarity measure, the decision to essentially binarize a continuous measure, the thresholds, the way of finding the prototype).  The advantage that your package allows is a kind of arbitrary transparency.  It provides a procedure that makes a choice, effectively tying the analysts hands so that they can claim they didn’t search over results (of course, in practice they could just search over parameters of your function as well).
