### 2020-06-17 JonasRieger/ldaPrototype#6 @bstewart
Defaults and Documentation
Regarding the JOSS review, I'd recommend documenting the default parameters passed to `lda.collapsed.gibbs.sampler` in the help files for the functions that call it. This is particularly important for those which don't have defaults or have different defaults in the original package.

It is a stylistic choice but I'd also give some consideration to removing the default for `K`.  Users rarely change defaults and I think a reason that other packages don't offer a default for `K` is a way of signaling that it is something that the user really has to engage with.
