### 2020-06-17 #2181 @bstewart
Hi all, my apologies for the long delay in my review.  I have posted a few issues in the target repo around the [paper](https://github.com/JonasRieger/ldaPrototype/issues/5), the need to [document defaults](https://github.com/JonasRieger/ldaPrototype/issues/6) and a small issue in the [README](https://github.com/JonasRieger/ldaPrototype/issues/4).

I have a broader philosophical question about the purpose of this package.  The implicit theory put forward by the author is that the best model to use is one that is broadly representative of other solutions.  But LDA is an optimization problem.  The reason we get different answers for different seeds is because that optimization problem is sufficiently hard to solve that we get stuck in local optima (a hopefully accessible introduction is here).  Under the optimization perspective ,the best solution is the one that maximizes the objective function.  So if we have a set of better and worse solutions why would we want one in the middle? This problem is made more vexing by the fact that certain decisions in the underlying `lda` package almost surely make this problem worse.  For example, it isn’t possible to retain anything more than the final draw of the Gibbs sampler for the topic-word counts although results would almost certainly stabilize (at least somewhat!) if you averaged over many draws.

I think this is an addressable concern.  The central advantage of the proposed procedure from where I sit is a kind of transparency. It suggests at least that of various solutions which are presumably locally optimal, this one shares some similarities with all of them.

Having not reviewed for this journal, I’m not really sure whether or not this last point is really in scope.  The package does what it says it does, but arguably this speaks to the question about the statement of need.  While I think a statement about why this is an attractive strategy in the paper would be preferred, I leave it to the editorial team (@karthik?) to make a judgment about whether such a discussion is relevant for publication here. I'm happy to remove this if it is not the kind of feedback you were looking for.

I'm running some longer tests of the functions which will take another few hours and then will report back on the last few bits.
