### 2020-06-17 #2181 @karthik
> Having not reviewed for this journal, I’m not really sure whether or not this last point is really in scope

I think this is useful and it would be good for the authors to address it, even briefly in the paper.
