### 2020-06-17 #2181 @bstewart
Okay the last thing I wanted to check finished running.  A core claim here is that the estimation is more stable under the proposed procedure.  I wanted to examine the evidence for that claim.  I ran the default dataset for 100 seeds with 100 topics (essentially all default settings).  I then did it again with a different set of seeds.  The question is how similar are the two sets of results?

After aligning the topics, here are the first three.

> Topic 1:
> said, dlrs, company, mln, pct, stock, inc
> said, dlrs, company, shares, mln, inc, common
> Topic 2:
> computer, technology, software, programmers, chinese, employs, francisco
> software, programmers, springboard, computer, applications, ever, lisp
> Topic 3:
> tower, washington, secretary, last, arms, criticism, state
> since, state, widespread, commission, made, william, thought

The first topic seems pretty well aligned, the next two are less so.  The rest of the first 10 topics are at the very bottom of this message (which are all I have examined).  The next relevant question is whether the two prototypes are more similar to each other than any other pair of models.  There is a question about the right measure of similar, so here I consider two, the correlation between the normalized and the unnormalized topic matrix (the latter being a little bit easier because you get some correlation just out of raw word frequencies).

![image](https://user-images.githubusercontent.com/577337/84960539-df579800-b0cf-11ea-8cee-1931cc2c645b.png)

The plot above shows the correlations of the prototypes with a red dashed line and a random sample of pairwise model comparisons as a null distribution.  We can see that while it is definitely more stable than random pairs, it isn't dramatically more stable.

There are plenty of things to criticize about this procedure.  It was a first cut at investigating what I took to be an empirical claim of the paper.  For one thing, this may be simply because the dataset considered here is relatively small and idiosyncratic.  However, evidence like this feels relevant to the long term project.

Complete code to regenerate these results is at the very bottom.  Note that while the LDA models run quickly the prototyping can take a couple of hours.

> Topic 4:
 > orders, economy, midland, took, laws, adversely, caused
 > orders, fed, decline, capital, economists, goods, wyss
> Topic 5:
 > said, january, december, year, pct, compared, rose
 > said, january, compared, december, new, pct, year
> Topic 6:
 > network, texas, behalf, comptroller, county, create, guaranteed
 > unit, capital, sources, power, made, alternate, bear
> Topic 7:
 > market, delay, likely, stocks, current, news, well
 > stock, banking, longer, market, delay, brazilian, equity
> Topic 8:
 > said, banks, interest, week, will, approval, terms
 > said, inc, reuter, first, industries, report, defense
> Topic 9:
 > inc, industries, reuter, said, burlington, cofab, petrochemical
 > interest, banks, bank, reuter, request, shock, afternoon
> Topic 10:
 > lion, mortgage, red, expects, partnership, covering, hotels
 > lion, red, limited, previously, bb, computerrelated, covering

```
#you will need the clue package for this

#run two default runs on the package with non-overlapping seeds
#we search over 100 solutions for the prototype here
reps_1 = LDARep(docs = reuters_docs, vocab = reuters_vocab,
              n = 100, seeds = 1:100)
reps_2 = LDARep(docs = reuters_docs, vocab = reuters_vocab,
              n = 100, seeds = 101:200)
#we extract the prototypes
p1 <- getPrototype(reps_1)
p2 <- getPrototype(reps_2)
#align the prototypes using the innerproducts of the beta matrix
align_it <- function(m1,m2) {
  x <- tcrossprod(getLDA(m1)$topics,getLDA(m2)$topics)
  clue::solve_LSAP(x, maximum = TRUE)
}
align <- align_it(p1,p2)
#compare the corresponding top words
words1 <- tosca::topWords(getTopics(getLDA(p2)), 7)[,align[1:10]]
words2 <- tosca::topWords(getTopics(getLDA(p1)), 7)[,1:10]
#print solutions
for(i in 1:10) {
  cat(sprintf("> Topic %i: \n > %s \n > %s \n", i,
              paste0(words1[,i],collapse=", "),
              paste0(words2[,i],collapse=", ")))
}
#check the correlations
b1 <- getTopics(getLDA(p2))[align,]
b2 <- getTopics(getLDA(p1))
c1 <- cor(c(b2),c(b1)) #without normalization
c2 <- cor(c(b2/rowSums(b2)),c(b1/rowSums(b1))) #with normalization

align_it <- function(m1,m2) {
  x <- tcrossprod(m1$topics,m2$topics)
  clue::solve_LSAP(x, maximum = TRUE)
}
set.seed(3253248)
grid <- cbind(sample.int(100,500,replace=TRUE),sample.int(100,500,replace=TRUE))
no_norm <- vector(mode="numeric",length=100)
norm <- vector(mode="numeric",length=100)
for(i in 1:nrow(grid)) {
  j <- grid[i,1]
  k <- grid[i,2]
  align <- align_it(reps_1$lda[[k]],reps_2$lda[[j]])
  b1 <- getTopics(reps_2$lda[[j]])[align,]
  b2 <- getTopics(reps_1$lda[[k]])
  no_norm[i] <- cor(c(b2),c(b1)) #without normalization
  norm[i] <- cor(c(b2/rowSums(b2)),c(b1/rowSums(b1))) #with normalization
}

par(mfrow=c(1,2))
hist(norm,main ="Normalizing",xlab="Rho",xlim=c(.2,.8))
abline(v=c2,lty=2,col="red",lwd=3)
hist(no_norm,main="No Normalizing",xlab="Rho",xlim=c(.2,.8))
abline(v=c1,lty=2,col="red",lwd=3)
```
