### 2020-06-18 JonasRieger/ldaPrototype#4 @JonasRieger
Fixed it, thank you!
edit: https://github.com/openjournals/joss-reviews/issues/2181
