### 2020-06-18 JonasRieger/ldaPrototype#6 @JonasRieger
Fixed it, thank you for your comments. I am with you that `K` should not have a default!
edit: https://github.com/openjournals/joss-reviews/issues/2181
