+++ paper/paper.md
@@ -23,12 +23,12 @@ affiliations:
 
 # Summary
 
-Topic Modeling [@blei2012] is one of the biggest subjects in the field of text data analysis. Here, the Latent Dirichlet Allocation [@blei2003] takes a special position. A large part of the analysis is based on this model (LDA). The LDA method has a far-reaching disadvantage. Random initialization and conditional reassignments within the iterative process of the Gibbs sampler [@griffiths2004] can result in fundamentally different models when executed several times on the same data and with identical parameter sets. This fact greatly limits the scientific reproducibility.
+Topic Modeling [@blei2012] is one of the biggest subjects in the field of text data analysis. Here, the Latent Dirichlet Allocation [@blei2003] takes a special position. A large part of scientific text data analyses are based on this model (LDA). The LDA method has a far-reaching disadvantage. Random initialization and conditional reassignments within the iterative process of the Gibbs sampler [@griffiths2004] can result in fundamentally different models when executed several times on the same data and with identical parameter sets. This fact greatly limits the scientific reproducibility.
 
-Up to now, the so-called eye-balling method has been used in practice to select suitable results. From a set of models, subjective decisions are made to select the model that seems to fit the data best. This contradicts basically good scientific practice. A different method of objective and automated selection has also become established. A model from a set of LDAs can be determined using the perplexity on held-out data. The [`R`](https://www.r-project.org/) [@R] package [`topicmodels`](https://CRAN.R-project.org/package=topicmodels) [@topicmodels] provides a workflow for this procedure. However, @chang2009 were able to show that this selection mechanism does not correspond to the human perception of a well-adapted model of text data. Instead, the authors propose a so-called intruder procedure based on human codings. The corresponding methodology is implemented in the package [`tosca`](https://github.com/Docma-TU/tosca) [@tosca].
+Up to now, the so-called eye-balling method has been used in practice to select suitable results. From a set of models, subjective decisions are made to select the model that seems to fit the data best or, in the worst case, the result that best supports one's hypothesis is chosen. This contradicts basically good scientific practice. A different method of objective and automated selection has also become established. A model from a set of LDAs can be determined optimizing the log-likelihood using the perplexity on held-out data. The [`R`](https://www.r-project.org/) [@R] package [`topicmodels`](https://CRAN.R-project.org/package=topicmodels) [@topicmodels] provides a workflow for this procedure. However, @chang2009 were able to show that this selection mechanism by optimizing likelihood-based measures does not correspond to the human perception of a well-adapted model of text data. Instead, the authors propose a so-called intruder procedure based on human codings. The corresponding methodology is implemented in the package [`tosca`](https://github.com/Docma-TU/tosca) [@tosca].
 
-The [`R`](https://www.r-project.org/) package [`ldaPrototype`](https://github.com/JonasRieger/ldaPrototype) on the other hand determines a prototypical LDA by automated selection from a set of LDAs. This is done using a typical statistical procedure. For a given combination of parameters, a number of models is calculated (usually about 100), from which that LDA is determined that is most similar to all other LDAs from a set of models. For this purpose pairwise model similarities are calculated using the S-CLOP measure (**S**imilarity of Multiple Sets by **C**lustering with **Lo**cal **P**runing), which can be determined by a clustering procedure of the individual topic units based on topic similarities of the two LDA results considered. The package offers visualization possibilities for comparisons of LDA models based on the clustering of the associated topics. Furthermore, the package supports the repetition of the modeling procedure of the LDA by a simple calculation of the repeated LDA runs.
[…]
