+++ paper/paper.bib
@@ -54,6 +54,47 @@
     doi = {10.18637/jss.v040.i13},
   }
   
+  @Manual{mallet,
+    title = {{mallet}: {A} wrapper around the {Java} machine learning tool {MALLET}},
+    author = {David Mimno},
+    year = {2013},
+    note = {R package version 1.0},
+    url = {https://CRAN.R-project.org/package=mallet},
+  }
+  
+  @Article{quanteda,
+    title = {{quanteda}: {An} {R} package for the quantitative analysis of textual data},
+    journal = {Journal of Open Source Software},
+    author = {Kenneth Benoit and Kohei Watanabe and Haiyan Wang and Paul Nulty and Adam Obeng and Stefan Müller and Akitaka Matsuo},
+    doi = {10.21105/joss.00774},
+    url = {https://quanteda.io},
+    volume = {3},
+    number = {30},
+    pages = {774},
+    year = {2018},
+  }
+  
+  @Article{stm,
+    title = {{stm}: An {R} Package for Structural Topic Models},
+    author = {Margaret E. Roberts and Brandon M. Stewart and Dustin Tingley},
+    journal = {Journal of Statistical Software},
+    year = {2019},
+    volume = {91},
+    number = {2},
+    pages = {1--40},
+    doi = {10.18637/jss.v091.i02},
+  }
+  
+@misc{rieger2020,
+    title={{Improving Reliability of Latent Dirichlet Allocation by Assessing Its Stability Using Clustering Techniques on Replicated Runs}},
+    author={Jonas Rieger and Lars Koppers and Carsten Jentsch and J\"org Rahnen\"uhrer},
+    year={2020},
+    eprint={2003.04980},
+    archivePrefix={arXiv},
+    primaryClass={cs.CL},
+    url={https://arxiv.org/abs/2003.04980}
+}  
+  
   @article{griffiths2004,
 	author = {Griffiths, Thomas L. and Steyvers, Mark},
 	title = {Finding scientific topics},
