+++ paper/paper.bib
@@ -75,7 +75,7 @@
   }
   
   @Article{stm,
-    title = {{stm}: An {R} Package for Structural Topic Models},
+    title = {{stm}: {An} {R} {Package} for {Structural} {Topic} {Models}},
     author = {Margaret E. Roberts and Brandon M. Stewart and Dustin Tingley},
     journal = {Journal of Statistical Software},
     year = {2019},
