### 2020-06-18 b6748a0
update paper averaging
