+++ paper/paper.bib
@@ -84,7 +84,19 @@
     pages = {1--40},
     doi = {10.18637/jss.v091.i02},
   }
-  
+
+@inproceedings{nguyen2014,
+    title = {{Sometimes Average is Best: The Importance of Averaging for Prediction using {MCMC} Inference in Topic Modeling}},
+    author = "Nguyen, Viet-An  and
+      Boyd-Graber, Jordan  and
+      Resnik, Philip",
+    booktitle = "Proceedings of the 2014 EMNLP-Conference",
+    year = "2014",
+    publisher = "ACL",
+    doi = "10.3115/v1/D14-1182",
+    pages = "1752--1757",
+}
+
 @misc{rieger2020,
     title={{Improving Reliability of Latent Dirichlet Allocation by Assessing Its Stability Using Clustering Techniques on Replicated Runs}},
     author={Jonas Rieger and Lars Koppers and Carsten Jentsch and J\"org Rahnen\"uhrer},
