+++ paper/paper.bib
@@ -90,7 +90,7 @@
     author = "Nguyen, Viet-An  and
       Boyd-Graber, Jordan  and
       Resnik, Philip",
-    booktitle = "Proceedings of the 2014 EMNLP-Conference",
+    booktitle = "Proceedings of the 2014 {EMNLP-Conference}",
     year = "2014",
     publisher = "ACL",
     doi = "10.3115/v1/D14-1182",
