### 2020-06-18 JonasRieger/ldaPrototype#5 @JonasRieger
Thank you!
- inserted mallet (and also stm and quanteda),
- improved wording,
- inserted more background information on different selection and averaging mechanisms.

edit: https://github.com/openjournals/joss-reviews/issues/2181
