### 2020-06-18 #2181 @JonasRieger
Thanks for all the feedback. I updated the default parametrization (no default for K) and fixed the bug in the readMe. I also improved the mentioned points in the paper. I hope this way the contribution will be clearer.
