### 2020-06-18 #2181 @bstewart
@JonasRieger - No problem, happy to help.
I still disagree with the characterization of two things in the draft.  I don't feel like these things need to change, but let me try to convince you.

1. I still think the language "This contradicts basically good scientific practice" is too strong.  Referencing the latter part of the sentence seems fine to me (i.e. obviously searching to fit your hypothesis is bad), but I see no reason that subjective assessments of model fit are any worse than a measure of model fit based on average similarity to a sample of local posterior modes (your approach).

2.In the Nguyen et al (2014), I don't think the averaging is a weakness.  Honestly, from a Bayesian perspective it is incredibly strange that standard practice isn't to do some kind of averaging.  What does one lose by abandoning a single sample from the posterior?  There is a potential problem with label switching if the chain is run long enough for sure, but honestly that is technically a problem with the single draws as well because the sampler is collapsed.

Anyway I don't feel like either of these should hold you up- just making another attempt to convince you.

@karthik @arfon I think I'm all set unless you need something else from me.
