### 2020-06-18 #2181 @JonasRieger
The statement actually refers mainly to the method of consciously choosing such models that correspond to the hypothesis. I have changed the specific wording to make this clear. @whedon generate pdf

The weakness of averaging is merely that not all facets of the model can be considered, or, as the case may be, it may become significantly more complex. The token specific assignments cannot be specified concretely, but at most probabilistically. This could lead to problems for some users who perform (partial) analyses on these topic assignments.

I am very interested in comparing the differences of the different selection and averaging algorithms. For this purpose I plan to do a qualitative study in the near future. However, I do not see any method as superior to the others by now.

Thank you again for your feedback and input. In fact, very helpful!
