### 2020-06-18 #2181 @bstewart
Glad you found it helpful.  Re: averaging- I take your point that the assignments are not concrete, but in fairness their concreteness is an illusion.  What you call concrete is just the particular realization of the Gibbs sampler on the draw you happened to stop things.

In my field of study (and most I know of), people are either examining theta (the documents by topics) or beta (words by topics). Both of these have the same form under averaging (values on the simplex).
