+++ man/ldaPrototype-package.Rd
@@ -59,6 +59,16 @@ For bug reports and feature requests please use the issue tracker:
 determines their Prototype.
 }
 
+\references{
+Rieger, Jonas, Lars Koppers, Carsten Jentsch and Jörg Rahnenführer (2020).
+"Improving Reliability of Latent Dirichlet Allocation by Assessing Its Stability using Clustering Techniques on Replicated Runs."
+arXiv 2003.04980, URL \url{https://arxiv.org/abs/2003.04980}.
+
+Rieger, Jonas, Jörg Rahnenführer and Carsten Jentsch (2020).
+"Improving Latent Dirichlet Allocation: On Reliability of the Novel Method LDAPrototype."
+In: \emph{Natural Language Processing and Information Systems, NLDB 2020.} LNCS 12089, pp. 118--125,
+DOI 10.1007/978-3-030-51310-8_11, URL \url{https://doi.org/10.1007/978-3-030-51310-8_11}.
+}
 \seealso{
 Useful links:
 \itemize{
