+++ man/ldaPrototype-package.Rd
@@ -60,6 +60,10 @@ determines their Prototype.
 }
 
 \references{
+Rieger, Jonas (2020). ldaPrototype: A method in R to get a Prototype of multiple Latent
+Dirichlet Allocations. Journal of Open Source Software, \bold{X}(xx), y,
+DOI 10.21105/joss.02181, URL \url{https://doi.org/10.21105/joss.02181}.
+
 Rieger, Jonas, Lars Koppers, Carsten Jentsch and Jörg Rahnenführer (2020).
 "Improving Reliability of Latent Dirichlet Allocation by Assessing Its Stability using Clustering Techniques on Replicated Runs."
 arXiv 2003.04980, URL \url{https://arxiv.org/abs/2003.04980}.
