+++ DESCRIPTION
@@ -28,5 +28,5 @@ Suggests:
 	RColorBrewer (>= 1.1-2),
 	testthat,
 	tosca
-RoxygenNote: 7.0.2
+RoxygenNote: 7.1.1
 LazyData: true
