### 2020-07-15 #2181 @JonasRieger
off course: 10.5281/zenodo.3945836 is the DOI of today's archive/version
10.5281/zenodo.3597978 is the concept DOI
