### 2020-07-15 #2181 @arfon
☝️ @karthik - running `@whedon accept` will add the appropriate label (`recommend-accept`) for the EiC on rotation to pick the paper up (this isn't me this week).
