+++ paper/paper.bib
@@ -99,7 +99,7 @@
 
 @misc{rieger2020,
     title={{Improving Reliability of Latent Dirichlet Allocation by Assessing Its Stability Using Clustering Techniques on Replicated Runs}},
-    author={Jonas Rieger and Lars Koppers and Carsten Jentsch and J\"org Rahnen\"uhrer},
+    author={Jonas Rieger and Lars Koppers and Carsten Jentsch and J\"org Rahnenf\"uhrer},
     year={2020},
     eprint={2003.04980},
     archivePrefix={arXiv},
