### 2020-07-15 #2181 @labarba
@JonasRieger -- please update the title metadata of your Zenodo deposit to match the JOSS paper title
