### 2020-07-15 #2181 @JonasRieger
@labarba I confirm that the version here should be updated to `0.2.0`. I also updated the title of the Zenodo deposit.
